+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [0.3.0] - 2020-09-16
+
 ### Changed
 
 - Made some performance tweaks to decompression.
@@ -31,6 +33,7 @@ and this project adheres to
 
 - Yaz0 decompression.
 
-[unreleased]: https://github.com/jordanbtucker/oead/compare/v0.2.0...HEAD
+[unreleased]: https://github.com/jordanbtucker/oead/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/jordanbtucker/oead/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/jordanbtucker/oead/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/jordanbtucker/oead/releases/tag/v0.1.0